`SOneTo` subtypes `AbstractOneTo` (#1302)
On a recent nightly (v"1.13.0-DEV.347"), `Base` has added
`AbstractOneTo` (https://github.com/JuliaLang/julia/pull/56902), and in
this PR we change `SOneTo` to subtype `AbstractOneTo` when this is
available. This makes combinations of `SOneTo`, `Base.OneTo` and
`Integer`s dispatch to `Base` methods, and we don't need to disambiguate
these in this package.

Fixes https://github.com/JuliaArrays/StaticArrays.jl/issues/1248. After
this, `StaticArrays` only needs to define `similar` methods for `Tuple`s
of `SOneTo` as indices.

---------

Co-authored-by: Neven Sajko <4944410+nsajko@users.noreply.github.com>
Co-authored-by: Mateusz Baran <mateuszbaran89@gmail.com>